==> notebooks/00_data_cleaning.ipynb <==
# Bigfive Exploration

import sys
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

print("Python version:", sys.version)
print("Pandas version:", pd.__version__)
print("Numpy version:", np.__version__)
print("Seaborn version:", sns.__version__)

#Import dataset
df = pd.read_csv("../data/raw/raw_data.csv", sep='\t')

### Data Overview

df.head()

df.shape

## Data Cleaning

#### Keep unique participants by removing duplicates by IP address


#Multiple IP counts could be the same person doing the test multiple times  
df[df['IPC'] != 1].shape

#Keep unique participants by removing duplicates by IP address
df_clean = df[df['IPC'] == 1]
df_clean.shape

#### Drop rows with any missing values in the 50 questions


#Drop rows with any missing values in the 50 questions
# Check total missing per question
import re

item_cols = [col for col in df_clean.columns if re.match(r'^(EXT|EST|AGR|CSN|OPN)\d+$', col)]
print(item_cols)
print(len(item_cols))  # Should be exactly 50


#### Drop rows with any invalid responses (score outside of 1-5) 


print(df_clean[item_cols].min().min())
print(df_clean[item_cols].max().max())


#Drop rows with any invalid responses (score outside of 1-5) 
invalid_score = (df_clean[item_cols] < 1) | (df_clean[item_cols] > 5)
print("Total invalid entries:", invalid_score.sum().sum())
rows_with_invalid = invalid_score.any(axis=1)
print("Rows with at least one invalid entry:", rows_with_invalid.sum())

df_clean = df_clean[~rows_with_invalid]
print(df_clean.shape)

#### Drop unrealistic testelapse values

#testelapse = the time(s) spent answering the 50 questions
#Finishing 50 questions under 20 seconds is near-impossible -> could be bots
df_clean = df_clean[df_clean['testelapse'] >= 20]
print(df_clean.shape)

#### Drop rows with introelapse == 0

#introelapse = time spent on the landing page -> shouldn't be 0
df_clean = df_clean[df_clean['introelapse'] > 0]
print(df_clean.shape)

#### Filter by Country [Optional]

#Filtering by Country lines for possible future use
#Selecting only Canadian entries
df_CA = df_clean[df_clean['country'].isin(['CA'])]
print(df_CA['country'].value_counts())
print(df_CA.shape)

##### Calculate mean score for each trait


# Calculate mean score for each trait
df_clean['Extraversion'] = df_clean[[f'EXT{i}' for i in range(1, 11)]].mean(axis=1)
df_clean['Emotional Stability'] = df_clean[[f'EST{i}' for i in range(1, 11)]].mean(axis=1)
df_clean['Agreeableness'] = df_clean[[f'AGR{i}' for i in range(1, 11)]].mean(axis=1)
df_clean['Conscientiousness'] = df_clean[[f'CSN{i}' for i in range(1, 11)]].mean(axis=1)
df_clean['Openness'] = df_clean[[f'OPN{i}' for i in range(1, 11)]].mean(axis=1)

df_clean[['Extraversion', 'Emotional Stability', 'Agreeableness', 'Conscientiousness', 'Openness']].head()

df_clean[['Extraversion', 'Emotional Stability', 'Agreeableness', 'Conscientiousness', 'Openness']].describe()

#### Reverse Coding Emotional Stability to Neuroticism


est_items = [col for col in df_clean.columns if col.startswith('EST') and not col.endswith('_E')]
est_items

#Reverse coding emotional stability items to neuroticism items
for i in range(1, 11):
    df_clean[f'NRT{i}'] = 6 - df_clean[f'EST{i}']

#Checking if the NRT columns are created 
nrt_items = [col for col in df_clean.columns if col.startswith('NRT')]
nrt_items

#Calculate mean score for Neuroticism
df_clean['Neuroticism'] = df_clean[[f'NRT{i}' for i in range(1, 11)]].mean(axis=1)


#Comparing EST and NRT
df_clean[['Neuroticism', 'Emotional Stability']]

#### Saving Cleaned Data in CSV

df_clean.to_csv('../data/cleaned/cleaned_bigfive_v1.csv', index=False)
